Refactor: Add handlers for extension related events
