+++ src/handlers/extensions/extensionAddedToNetwork.ts
@@ -1,0 +1,6 @@
+import { ContractEvent } from '~/types';
+import { writeExtensionVersionFromEvent } from '~/utils';
+
+export default async (event: ContractEvent): Promise<void> => {
+  await writeExtensionVersionFromEvent(event);
+};
