+++ src/handlers/extensions/extensionUninstalled.ts
@@ -1,0 +1,6 @@
+import { ContractEvent } from '~/types';
+import { deleteExtensionFromEvent } from '~/utils';
+
+export default async (event: ContractEvent): Promise<void> => {
+  await deleteExtensionFromEvent(event);
+};
