+++ src/handlers/extensions/index.ts
@@ -1,0 +1,6 @@
+export { default as handleExtensionAddedToNetwork } from './extensionAddedToNetwork';
+export { default as handleExtensionInstalled } from './extensionInstalled';
+export { default as handleExtensionUninstalled } from './extensionUninstalled';
+export { default as handleExtensionDeprecated } from './extensionDeprecated';
+export { default as handleExtensionUpgraded } from './extensionUpgraded';
+export { default as handleExtensionInitialised } from './extensionInitialised';
