+++ src/handlers/index.ts
@@ -1,2 +1,3 @@
 export * from './colonies';
 export * from './actions';
+export * from './extensions';
